+++ PKHeX.Core/Legality/Encounters/Generator/Possible/EncounterPossible4.cs
@@ -222,7 +222,7 @@ public record struct EncounterPossible4(EvoCriteria[] Chain, EncounterTypeGroup 
         for (; Index < db.Length;)
         {
             var enc = db[Index++];
-            if (!enc.CanBeReceivedByVersion(Version))
+            if (!enc.CanBeReceivedByVersion((int)Version))
                 continue;
             foreach (var evo in Chain)
             {
